Add option to return TCAV results as a proto
Previously, TCAV.run() returned its results as a list of dicts, where
each dict contained a variety of metrics.

Now, an arg to TCAV.run(), return_proto, can be set to True to return a
list of protos instead, where the proto format is defined in
results.proto and contains the same fields as the previous results dict.

A method to convert the dict to proto format is in utils.py.

In addition, return two new metrics - negative_concept and
cav_accuracies.